Up to line 11.
Just a quick summary of the elements:

`<g>` is a group. It's purpose was originally to group what a particular character does, but I think I might decide it's there for convenience, and basically lets you define attributes that are inherited by the children, if applicable. Depends what's easier to XSLT, I guess. (I'm thinking a template param containing an attribute list, if that's possible?)

`<loc>` is a location. Its presence makes me kind of think I should actually have a `<scene>` root element, instead of that "Shakespeare says" nonsense that I've got at the moment.

`<enter>` says that characters have entered. It can be used both with `<exit>` and `<exeunt>` syntax, if that's easy to XPath, but if not I'll defer to `<exeunt>` syntax.

`<c>` is a character reference. If it's empty, the content should be assumed to be the contents of the `name` attribute; otherwise it should be the content.

`<v>` is a verb. It can have a dative, the `to` attribute. If it's got text content, treat it as though that's wrapped in a single `<t>` element.

`<a>` is an anchor, with different semantics to the HTML anchor but a similar purpose. It doesn't link to things.

`<t>` is like HTML's `<p>`, ish, except it is more like `<li>`. You know what I mean, me!

`<q>` is identical to HTML's `<q>`. However, it belongs in this namespace because it's used a lot. (Can you import something from one namespace to another? That'd be great!)

`<exeunt>` is like `<exit>`, except for more than one person. They're listed in `<c>` tags inside this element.

`<exit>` is like `<exeunt>`, except for one person. As such, it uses the `name` attribute directly.

That's about it! Make sure you remember about this commit and don't spend too long making the schema; I've just done the hard work for you.